feat: Add faviconDownloader and failure icon #2
- Add failure icon
- Add FaviconDownloaderProtocol
Write FaviconDownloaderProtocol, to be the Protocol that used to any FaviconDownloader or FaviconDownloaderMock in testing
To be implemented by FaviconDownloader and encapsulate the downloading implementation of the Favicon.
Remove or replace FaviconFinder any time easily without any side effects

- Add FaviconDownloader that implement FaviconDownloaderProtocol which do the following functions:

check if the website url is valid to download the Favicon first before sending download request.
use FaviconFinder to fetch the Favicon.
run completion when with the downloaded FaviconFinder or the error that happen.
the error could be one of the 2 following errors:
Downloading failed for the one of the following reasons
1. no internet
2. emptyData
3. failedToParseHTML
4. failedToFindFavicon
5. failedToDownloadFavicon
6. emptyFavicon
7. notValidImage
the website url is not valid